requirements.txt
drf-yasg >= 1.20.0
psycopg2 >= 2.8.6
djangorestframework >= 3.12.4
django-environ >= 0.8.1